Fix Java interop serialization issues  (#19)
* Fixes steps serialization when called from Java

It's not possible to call reified inline Kotlin methods from Java
so steps return types will have to be explicitly passed.
Without `reified inline` generic types are lost by at runtime and
cannot be used anymore for deserialization purposes.

https://stackoverflow.com/questions/42741780/how-can-i-call-kotlin-methods-with-reified-generics-from-java/42742119#42742119

https://docs.oracle.com/javase/tutorial/java/generics/erasure.html

* Fix for the empty object serialization issue

For now steps return types will need to be:
    - Public classes
    - Having public `JsonProperty`s
so that getters can be called from the `inngest-core` package.

It's quite limiting I think. Getters do not work as expected with
Klaxon unless it's declared in Kotlin.

But we'll try to improve it later on.