Setup workflow to upload to Maven central (#49)
Update the SDK core's `build.gradle.kts` to publish to a local Maven repo (which is just basically a file path), and bundle that to send it over to Maven Central's publisher API.

The new Maven Central don't have integration with gradle atm, so unfortunately, we'll just have to automate it in a way simulating a manual upload.
https://central.sonatype.org/publish/publish-portal-api/

The `Release` github workflow is set to trigger when there are changes made to each project's `VERSION` file. Which assumes that we'll want to publish a new version.

---------
Co-authored-by: Darwin D Wu <darwin67@users.noreply.github.com>