Implement an interface for functions that need to handle failures.
I took a more OOP approach for this feature which is slightly different
than the other SDKs but we can discuss this further in the PR review.

The `WithFailureHandler` interface has a single method `onFailure` that
is called when the function fails. An example of how to use this:

```kotlin
class MyFunction : InngestFunction(), WithFailureHandler {
    override fun execute(ctx: FunctionContext, step: Step): Any? {
        // Do something
    }

    override fun onFailure(ctx: FunctionContext, step: Step): Any? {
        // Handle failure
    }
}
```