Support try-catch for both `invoke` and `run`
If an exception is thrown in the `invoke` or `run` function, it will now
be caught and returned as a `StepError` exception. This will allow the
error to be propagated to the caller and handled appropriately.

Transforming the errors into a `StepError` exception is following the
Inngest SDK spec:
https://github.com/inngest/inngest/blob/main/docs/SDK_SPEC.md#522-memoizing-a-step
Inngest JS SDK behavior:
https://github.com/inngest/inngest-js/blob/4f91d9c302592ecc2228914469dd057ae148005b/packages/inngest/src/components/execution/v1.ts#L437-L443
Inngest documentation:
https://www.inngest.com/docs/features/inngest-functions/error-retries/inngest-errors#step-errors