Implement debounce configuration (#70)
Follows similar pattern to recently added `throttle`

The integration test timing could lead to flakes but I tried to leave
a generous buffer to minimize this without making the test too long.

I wanted to add a second integration test for the debounce `timeout`,
but configuring debounce with a `timeout` seems to result in the
function never getting executed on the latest dev server (0.29.6). I
verified this is true with the JS SDK as well.