Add tests for cloud mode introspection
It's a pain to mock the environment variables in the tests. I introduced
a different `system-stubs-jupiter` library to help with that because
`org.junitpioneer.jupiter` did not work as expected in a Spring Boot
environment.

Ideally I think we should have a mockable custom `Environment` interface
that we use throughout the application, instead of reaching for
`System.getenv()` directly.

The method for mocking that worked is described in this guide:
https://www.baeldung.com/java-system-stubs#environment-and-property-overrides-for-junit-5-spring-tests