Fix serialization of subclasses (#92)
* Add integration test for catching Exception instead of StepError

This is the reproduction of an error reported by a user. Because
StepError contains additional fields to a generic Exception,
deserialization fails.

* Add integration test for the general case of deserializing a subclass

This shows that the issue is not limited to StepError/Exception.

* Fix deserialization of subclasses

Previously we were deserializing step data based on the T provided by the compiler.
However, this could be lossy if the step data was actually a subclass of
T with additional fields and is semantically incorrect if we try to
deserialize it as the parent class.

To remedy this, we need to know what the actual class of the step data
was before it was memoized with the Inngest server. We are adding this
as an additional field `class` on the JSON when serializing, then
reading (and deleting) this field and using it with reflection to
deserialize using the proper class.

It might make sense to put the field `class` in some kind of
meta/extension section of the payload to avoid conflict, but it's highly
unlikely a field will legitimately be named class in the user code since
it is a reserved keyword in Java.

This strategy doesn't entirely work for schema evolution because it's
still possible the classes have changed to the point where
deserialization would fail. We should follow up here and probably have a
fallback mechanism of returning a generic hash if deserialization fails.

---------

Co-authored-by: Albert Chae <albertchae@users.noreply.github.com>
Co-authored-by: Riadh <22998716+KiKoS0@users.noreply.github.com>